Redução de linhas de código no menu
