+++ src/ui/UsuarioUI.java
@@ -1,0 +1,13 @@
+package ui;
+
+import java.util.Scanner;
+
+public class UsuarioUI {
+public static void main(String[] args) {
+	
+	Scanner sc = new Scanner(System.in);
+	Menu m = new Menu(sc);
+	
+	m.menuInicial();
+}
+}
